Adding a structure to the project
All the objects were being generated on the same
directory as the main, this is not moved to `include/`.

The external CUDA files were also moved to `include/`.

Because of the renaming, the Makefile was adapted to
fit this new structure.

The documentation was also placed under the `doc/`
directory.

Signed-off-by: Cristian Maureira-Fredes <cmaureirafredes@gmail.com>